Refactor Layout structure and enhance DXF handling
- Changed `tab_order` in `Layout` from `i16` to `i32` for better compatibility.
- Updated `SectionReader` to cast `tab_order` to `i32` when reading from DXF.
- Modified `SectionWriter` to cast `tab_order` back to `i16` when writing to DXF.
- Added new fields to `Layout` struct: `ucs_origin`, `ucs_x_axis`, `ucs_y_axis`, `elevation`, and `ucs_ortho_type` for improved UCS handling.
- Implemented a `model` constructor for `Layout` to facilitate model space creation.
- Enhanced tests for `Layout` to cover new fields and ensure proper functionality.
- Added comprehensive tests for critical non-graphical object writers, including dictionaries, XRecords, and PlotSettings.
- Ensured that all new features are covered by unit tests to maintain code quality and reliability.

+++ src/io/dxf/reader/section_reader.rs
@@ -1217,7 +1217,7 @@ impl<'a> SectionReader<'a> {
                 71 => {
                     // Tab order
                     if let Some(value) = pair.as_i16() {
-                        layout.tab_order = value;
+                        layout.tab_order = value as i32;
                     }
                 }
                 10 => {
